requirements.txt
flask
flask-sqlalchemy
requests
pymysql
python-dotenv
